// projects/vocabbook/src/types/vocabulary.ts
export interface WordData {
  표제어: string;
  품사: string;
  난이도: number;
  표제어뜻: string;
  영영정의: string;
  발음: string;  // Added this field
  동의어: string[];
  동의어뜻: string[];
  반의어: string[];
  반의어뜻: string[];
}

export interface ParsedWordData {
  words: WordData[];
  title: string;
}